requirements.txt
tensorflow 
opencv-python 
pandas 
numpy 
Pillow